Fix Neo4j connection when running in Docker
This change fixes the connection string to Neo4j when running
inside a Docker container. It also delays app startup until a
healthy Neo4j graph client can be constructed and execute a query
against the database. Lastly, it fixes a boolean bug when checking
that a system shouldn't exist before fetching it from ESI.